Internals now based on array of entries.
Reimplmented the filters internally to operate on an array of entries,
rather than a real feed. This prevents some nasty work-arounds for
the fact that XML::Atom doesn't allow manipulation of entries. It also
allows me to use Moose::Meta::Attribute::Native and standard list
manipulation routines, which simplifies the implementation.