Entries now get a source element added
When adding entries to a pipe, each entry now gets a source element
added to it (unless one already exists). This is so that the resulting
output does not lose information as to where the entry came from, and
carries with it information normally "inherited" from the feed.

Some information is still being lost, however, when the feeds are
converted from RSS to Atom. That's a different problem.

Making this work properly required fixing a bug in XML::Atom.
We now depend on XML::Atom 0.37 as a result.